add st_experimental_rerun, add test within e2e
